\name{RServerUtils-package}
\alias{RServerUtils-package}
\alias{RServerUtils}
\docType{package}
\title{
\packageTitle{RServerUtils}
}
\description{
\packageDescription{RServerUtils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RServerUtils}
\packageIndices{RServerUtils}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RServerUtils}

Maintainer: \packageMaintainer{RServerUtils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
